authentication added

+++ app/Http/Controllers/AuthController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\authRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
+use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
@@ -51,7 +52,7 @@ class AuthController extends Controller
                   "message"=> "email or password is incorrect"]
              );
        }
-      $token =  $user->createToken("auth_token",["read"])->plainTextToken;
+      $token =  $user->createToken("auth_token",["read-book",'update-book','delete-book','insert-book',"read-author"])->plainTextToken;
       return response()->json([
          "success"=> true,
          "user"=> new UserResource($user),
@@ -61,12 +62,27 @@ class AuthController extends Controller
      }
 
      public function logout(Request $request){
-       $request->user()->tokens()->delete();
-       return response()->json(
-         [
-            "data"=> "you are logged out"
-         ]
-       );
+     try{ 
+     if($request->user() && $request->user()->currentAccessToken()){
+           $request->user()->tokens()->delete();
+           return response()->json([
+            "message"=> "you are logged out"
+           ]);
+      }
+      }catch(Exception $e){
+         return response()->json([
+            "message"=> $e->getMessage(),
+         ]);
+      }
+      // return response()->json([
+      //    "message"=> "user has already been logout out",
+      // ]);
+      //  $request->user()->currentAccessToken()->delete();
+      //  return response()->json(
+      //    [
+      //       "data"=> "you are logged out"
+      //    ]
+      //  );
      }
